+++ x/ibcratelimit/simulation/operations.go
@@ -7,6 +7,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/module"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
+
 	simappparams "github.com/provenance-io/provenance/app/params"
 	"github.com/provenance-io/provenance/x/ibcratelimit"
 	"github.com/provenance-io/provenance/x/ibcratelimit/keeper"
@@ -31,7 +32,7 @@ func ProposalMsgs(simState module.SimulationState, k *keeper.Keeper) []simtypes.
 }
 
 func SimulatePropMsgUpdateParams(k *keeper.Keeper) simtypes.MsgSimulatorFn {
-	return func(r *rand.Rand, ctx sdk.Context, accs []simtypes.Account) sdk.Msg {
+	return func(r *rand.Rand, _ sdk.Context, _ []simtypes.Account) sdk.Msg {
 		// change it to a new random account.
 		raccs := simtypes.RandomAccounts(r, 1)
 		return ibcratelimit.NewMsgUpdateParamsRequest(k.GetAuthority(), raccs[0].Address.String())